separate package analyzer, musicInfo
